add tests for fetchFilteredData api endpoint

+++ apps/frontend/src/app/api/lib/utils.ts
@@ -257,8 +257,8 @@ export const searchQueryFieldSchema = z.array(
 
 export const fetchFilteredDataSchema = z.object({
   searchQuery: searchQueryFieldSchema,
-  pageSize: z.number().min(1).max(100),
-  page: z.number().min(1),
+  pageSize: z.number().int().min(1).max(100),
+  page: z.number().int().min(1),
 });
 
 // const partyCodes = [
